feat(navigation): 增强导航与调试功能 引入了导航数据的记录与回放功能，提升了系统对运动轨迹的跟踪与分析能力。 增加了 Flash 读取状态的重置功能，以确保数据处理的准确性。 在 UI 中新增了一个调试界面，用于实时监控里程（mileage）和保存索引（save index）等导航数据指标。 更新了导航逻辑，优化了里程计算的处理方式，确保在数据记录和回放过程中运行更加流畅。 完善了 LED 指示灯的 GPIO 集成，以便在发生导航事件时提供直观的视觉反馈。

+++ Blalance_wheel_leg_Feis_v1.4/project/code/small_driver_uart_control.c
@@ -2,12 +2,12 @@
 
 small_device_value_struct motor_value; // 定义通讯参数结构体
 
-int16 Left_Motor_Pwm = 0;  // 左电机输出pwm
-int16 Right_Motor_Pwm = 0; // 右电机输出pwm
-uint8 Motor_Switch = 1;    // 电机开关
-uint8 Motor_OpenFlag = 0;  // 电机上电标志位
-uint8 Left_Motor_Speed = 0; // 左电机速度
-uint8 Right_Motor_Speed = 0; // 右电机速度
+int16 Left_Motor_Pwm = 0;   // 左电机输出pwm
+int16 Right_Motor_Pwm = 0;  // 右电机输出pwm
+uint8 Motor_Switch = 1;     // 电机开关
+uint8 Motor_OpenFlag = 0;   // 电机上电标志位
+int16 Left_Motor_Speed = 0; // 左电机速度
+int16 Right_Motor_Speed = 0; // 右电机速度
 
 //-------------------------------------------------------------------------------------------------------------------
 // 函数简介     无刷驱动 串口接收回调函数
